# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Production-ready URL shortening service with Redis persistence, caching, rate limiting, and comprehensive management features.",
        "title": "Short URL Generator API",
        "termsOfService": "https://github.com/yourusername/short-url-generator",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/yourusername/short-url-generator/issues",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/user/analytics": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get comprehensive analytics for authenticated user including click trends, device breakdown, and top URLs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user analytics",
                "responses": {
                    "200": {
                        "description": "User analytics data",
                        "schema": {
                            "$ref": "#/definitions/model.UserAnalytics"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/user/url/{shortURL}/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed access logs for a specific short URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get URL access logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL",
                        "name": "shortURL",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Access logs",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "URL does not belong to user",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ActivityLog": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "Type of action (see ActivityType constants)",
                    "type": "string"
                },
                "details": {
                    "description": "Additional action-specific details",
                    "type": "object",
                    "additionalProperties": true
                },
                "ip": {
                    "description": "IP address of the user",
                    "type": "string"
                },
                "location": {
                    "description": "Optional: City, Country (from GeoIP)",
                    "type": "string"
                },
                "timestamp": {
                    "description": "When the action occurred",
                    "type": "string"
                },
                "userAgent": {
                    "description": "Browser/device user agent",
                    "type": "string"
                }
            }
        },
...
        "model.ErrorResponse": {
            "description": "Standard error response",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid URL format"
                }
            }
        },
...
        "model.TimeSeriesPoint": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "Date in \"YYYY-MM-DD\" format",
                    "type": "string"
                },
                "value": {
                    "description": "Number of clicks on this date",
                    "type": "integer"
                }
            }
        },
...
        "model.URLStats": {
            "type": "object",
            "properties": {
                "clicks": {
                    "description": "Total number of clicks",
                    "type": "integer"
                },
                "lastAccessed": {
                    "description": "Last access timestamp (ISO 8601)",
                    "type": "string"
                },
                "originalURL": {
                    "description": "Original destination URL",
                    "type": "string"
                },
                "shortURL": {
                    "description": "Short URL identifier",
                    "type": "string"
                }
            }
        },
...
        "model.UserAnalytics": {
            "type": "object",
            "properties": {
                "activeUrls": {
                    "description": "Number of active URLs",
                    "type": "integer"
                },
                "browserBreakdown": {
                    "description": "Clicks by browser",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "clicksByDay": {
                    "description": "Time-series data for clicks",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TimeSeriesPoint"
                    }
                },
                "deviceBreakdown": {
                    "description": "Clicks by device type (mobile/desktop/tablet)",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "recentActivity": {
                    "description": "Recent user activities",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ActivityLog"
                    }
                },
                "topUrls": {
                    "description": "Top performing URLs by clicks",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.URLStats"
                    }
                },
                "totalClicks": {
                    "description": "Total clicks across all URLs",
                    "type": "integer"
                },
                "totalUrls": {
                    "description": "Total number of URLs created",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Admin API key for accessing protected endpoints",
            "type": "apiKey",
            "name": "X-Admin-Key",
            "in": "header"
        }
    },
...
}